Mini boiler accepts input from the side
As it is right now, any pipe inputting fluids into mini boilers connected to Pressure Refineries will replace partly the lava or fire used by that machine. This change aims at fixing that by allowing fluids to be inserted either from the bottom or the side opposite to the machine being upgraded.