Fix alchemy tablet breaking after 10 days of existing
The alchemy tablet seizes to function after existing for 16777216 ticks. This happens because the angle field gets incremented by a float value instead of an int value. The number of ticks is the maximum number a 24 bit value can represent and after this it does not increment anymore. If this happens the value will forever be stuck at 16777216 and the condition `angle % 10 == 0` will never be true, thus preventing any recipe from ever finishing.